refactor: standardize terminology from "summarise" to "summarize" across components; add UserStatistics model
